Made further adjustments to tests and example based on reviewer feedback
