+++ Source/ViewControllers/BrickViewController.swift
@@ -8,14 +8,12 @@
 
 import UIKit
 
-#if os(iOS)
 /**
  Conform to this delegate on your preview view controller in order to enable UIKit Pop.
  */
 public protocol BrickViewControllerPreviewing: class {
     var sourceBrick: Brick? { get set }
 }
-#endif
 
 /// A BrickViewController is a UIViewController that contains a BrickCollectionView
 open class BrickViewController: UIViewController, UICollectionViewDelegate {
@@ -41,7 +39,10 @@ open class BrickViewController: UIViewController, UICollectionViewDelegate {
         }
         return collectionView
     }
-
+    
+    /// Previewing context, retained for unregistration in the event of a capability change
+    open internal(set) var previewingContext: UIViewControllerPreviewing?
+    
     #if os(iOS)
     // MARK: - Internal members
 
@@ -50,9 +51,6 @@ open class BrickViewController: UIViewController, UICollectionViewDelegate {
 
     /// Refresh action, if added
     open internal(set) var refreshAction: ((_ refreshControl: UIRefreshControl) -> Void)?
-
-    /// Previewing context, retained for unregistration in the event of a capability change
-    open internal(set) var previewingContext: UIViewControllerPreviewing?
     #endif
 
     // MARK: - Initializers
@@ -79,6 +77,11 @@ open class BrickViewController: UIViewController, UICollectionViewDelegate {
             registrationDataSource.registerBricks()
         }
     }
+    
+    /// Registration/unregistration happens here because the user can disable force touch in their system settings
+    open override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
+        handleTraitCollectionChange(traitCollection, previousTraitCollection)
+    }
 
     // MARK: - Private Methods
 
@@ -92,6 +95,19 @@ open class BrickViewController: UIViewController, UICollectionViewDelegate {
 
             self.collectionView = collectionView
             self.collectionView?.delegate = self
+        }
+    }
+    
+    /// On tvOS, this method accomplishes nothing, despite UIViewControllerPreviewing and its delegate protocol being available
+    func handleTraitCollectionChange(_ traitCollection: UITraitCollection, _ previousTraitCollection: UITraitCollection?) {
+        if previewingContext == nil && traitCollection.forceTouchCapability == .available {
+            // We use brickCollectionView as the sourceView for the previewingContext, otherwise the coordinates returned
+            // in UIViewControllerPreviewing#previewingContext(_:, viewControllerForLocation:) are not in the right system
+            previewingContext = registerForPreviewing(with: self, sourceView: brickCollectionView)
+        } else if let previewingContext = previewingContext,
+            previousTraitCollection?.forceTouchCapability == .available {
+            unregisterForPreviewing(withContext: previewingContext)
+            self.previewingContext = nil
         }
     }
 
@@ -115,7 +131,30 @@ open class BrickViewController: UIViewController, UICollectionViewDelegate {
     open func invalidateLayout() {
         brickCollectionView.invalidateBricks()
     }
+}
 
+// MARK: - UIViewControllerPreviewingDelegate
+extension BrickViewController: UIViewControllerPreviewingDelegate {
+    open func previewingContext(_ previewingContext: UIViewControllerPreviewing, viewControllerForLocation location: CGPoint) -> UIViewController? {
+        guard let indexPath = brickCollectionView.indexPathForItem(at: location) else {
+            return nil
+        }
+        let brick = brickCollectionView.brick(at: indexPath)
+        let viewController = brick.previewingDelegate?.previewViewController
+        if let previewing = viewController as? BrickViewControllerPreviewing {
+            previewing.sourceBrick = brick
+        }
+        return viewController
+    }
+    
+    open func previewingContext(_ previewingContext: UIViewControllerPreviewing, commit viewControllerToCommit: UIViewController) {
+        guard let previewViewController = viewControllerToCommit as? BrickViewControllerPreviewing else {
+            return
+        }
+        let sourceBrick = previewViewController.sourceBrick
+        let previewingDelegate = sourceBrick?.previewingDelegate
+        previewingDelegate?.commit(viewController: viewControllerToCommit)
+    }
 }
 
 #if os(iOS)
@@ -146,29 +185,12 @@ extension BrickViewController {
             self.refreshAction?(refreshControl)
         }
     }
-    
-    /// Registration/unregistration happens here because the user can disable force touch in their system settings
-    open override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
-        handleTraitCollectionChange(traitCollection, previousTraitCollection)
-    }
-    
-    func handleTraitCollectionChange(_ traitCollection: UITraitCollection, _ previousTraitCollection: UITraitCollection?) {
-        if previewingContext == nil && traitCollection.forceTouchCapability == .available {
-            /// We use brickCollectionView as the sourceView for the previewingContext, otherwise the coordinates returned
-            /// in UIViewControllerPreviewing#previewingContext(_:, viewControllerForLocation:) are not in the right system
-            previewingContext = registerForPreviewing(with: self, sourceView: brickCollectionView)
-        } else if let previewingContext = previewingContext,
-            previousTraitCollection?.forceTouchCapability == .available {
-            unregisterForPreviewing(withContext: previewingContext)
-            self.previewingContext = nil
-        }
-    }
 }
 #endif
 
 #if os(tvOS)
 extension BrickViewController {
-    
+
     open func collectionViewShouldUpdateFocusIn(context: UICollectionViewFocusUpdateContext) -> Bool {
         
         guard let nextIndex = context.nextFocusedIndexPath else {
